+++ packages/craftcms-legacy/axios/webpack.config.js
@@ -9,7 +9,7 @@ module.exports = getConfig({
   context: __dirname,
   config: {
     output: {
-      path: __dirname + '/../../../resources/legacy/axios/dist',
+      path: __dirname + '/../../../cms-assets/resources/legacy/axios/dist',
     },
     plugins: [
       new CopyWebpackPlugin({