+++ routes/actions.php
@@ -55,8 +55,6 @@ use CraftCms\Cms\Http\Controllers\Entries\ReassignEntriesModalController;
 use CraftCms\Cms\Http\Controllers\Entries\StoreEntryController;
 use CraftCms\Cms\Http\Controllers\FieldsController;
 use CraftCms\Cms\Http\Controllers\Gql\ApiController as GqlApiController;
-use CraftCms\Cms\Http\Controllers\Gql\SchemasController as GqlSchemasController;
-use CraftCms\Cms\Http\Controllers\Gql\TokensController as GqlTokensController;
 use CraftCms\Cms\Http\Controllers\IconController;
 use CraftCms\Cms\Http\Controllers\Import\ImportConfigController;
 use CraftCms\Cms\Http\Controllers\Import\ImportRunController;
@@ -73,9 +71,7 @@ use CraftCms\Cms\Http\Controllers\QueueController;
 use CraftCms\Cms\Http\Controllers\RelationalFieldsController;
 use CraftCms\Cms\Http\Controllers\Settings\EntryTypesController;
 use CraftCms\Cms\Http\Controllers\Settings\FilesystemsController;
-use CraftCms\Cms\Http\Controllers\Settings\ImageTransformsController;
 use CraftCms\Cms\Http\Controllers\Settings\SectionsController;
-use CraftCms\Cms\Http\Controllers\Settings\UserSettingsController;
 use CraftCms\Cms\Http\Controllers\Settings\VolumesController;
 use CraftCms\Cms\Http\Controllers\StructuresController;
 use CraftCms\Cms\Http\Controllers\Updates\UpdaterController;
@@ -342,24 +338,6 @@ Route::prefix($routes->cpActionTriggerRoutePrefix())->middleware(['craft.cp'])->
         // FindAndReplace
         Route::post('utilities/find-and-replace-perform-action', FindAndReplaceController::class);
 
-        // GraphQL
-        Route::middleware([RequireAdmin::class])->group(function () {
-            Route::post('graphql/generate-token', [GqlTokensController::class, 'generate']);
-
-            Route::middleware('password.confirm')->group(function () {
-                Route::post('graphql/save-token', [GqlTokensController::class, 'store']);
-                Route::post('graphql/fetch-token', [GqlTokensController::class, 'fetch']);
-            });
-        });
-
-        Route::middleware([RequireAdminChanges::class])->group(function () {
-
-            Route::middleware('password.confirm')->group(function () {
-                Route::post('graphql/save-schema', [GqlSchemasController::class, 'save']);
-                Route::post('graphql/save-public-schema', [GqlSchemasController::class, 'savePublic']);
-            });
-        });
-
         // Import
         Route::middleware('can:editImportConfigs')->group(function () {
             Route::post('import/configs/render-settings', [ImportConfigController::class, 'renderSettings']);
@@ -451,7 +429,6 @@ Route::prefix($routes->cpActionTriggerRoutePrefix())->middleware(['craft.cp'])->
         Route::middleware([RequireAdminChanges::class])->group(function () {
             Route::post('volumes/save-volume', [VolumesController::class, 'save']);
             Route::post('volumes/reorder-volumes', [VolumesController::class, 'reorder']);
-            Route::post('image-transforms/save', [ImageTransformsController::class, 'save']);
         });
 
         // Plugins
@@ -536,13 +513,6 @@ Route::prefix($routes->cpActionTriggerRoutePrefix())->middleware(['craft.cp'])->
         Route::post('users/remove-password-reset-requirement', [PasswordController::class, 'removeResetRequirement']);
         Route::post('users/verify-password', [PasswordController::class, 'verifyPassword']);
         Route::post('users/save-field-layout', SaveUsersFieldLayoutController::class);
-
-        // User settings
-        Route::middleware([
-            RequireAdminChanges::class,
-        ])->group(function () {
-            Route::post('user-settings/save-user-settings', [UserSettingsController::class, 'store']);
-        });
 
         // Pluginstore
         Route::middleware([